:memo: [Kubernetes] Update : K8S Resource Process Update
